+++ extras/examples/custom_port.rs
@@ -9,10 +9,11 @@
 
 use bevy::prelude::*;
 use bevy_brp_extras::BrpExtrasPlugin;
+use bevy_brp_extras::PortDisplay;
 
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_plugins(BrpExtrasPlugin::with_port(9000))
+        .add_plugins(BrpExtrasPlugin::with_port(9000).port_in_title(PortDisplay::Always))
         .run();
 }